Split out FamilyValues
