+++ families/Families.hs
@@ -9,19 +9,11 @@ import Data.Functor.Const (Const)
 import Data.Kind (Type)
 import Data.Map (Map)
 import Numeric.Natural (Natural)
-import Data.Proxy (Proxy)
-import Ledger (PubKey, Signature, SlotRange)
 
 class ValidatorScript s where
   type Currencies s :: [k]
   type Datum s    :: Type
   type Redeemer s :: Type
-
-type DatumSpecimen :: forall script -> Datum script -> Type
-type family DatumSpecimen s :: Datum s -> Type
-
-type RedeemerSpecimen :: forall script -> Redeemer script -> Type
-type family RedeemerSpecimen s :: Redeemer s -> Type
 
 type DApp :: fam -> Type
 type family DApp t
@@ -39,49 +31,10 @@ class Transaction (t :: familie) where
                  -> Type
   type Mints t = Const ()
 
-data TxSpecimen t = TxSpecimen {
-  txInputs :: Inputs t TxInputSpecimen WalletSpecimen,
-  txCollateral :: WalletSpecimen Collateral,
-  txOutputs :: Outputs t TxOutSpecimen WalletSpecimen,
-  txMint :: Mints t TxMintSpecimen,
-  txValidRange :: !SlotRange,
-  txFee :: Value '[ 'Ada ],
-  txSignatures :: Map PubKey Signature}
-
-type TxInputSpecimen :: forall (s :: script) -> Redeemer s -> Datum s -> [currency] -> Type
-data TxInputSpecimen s r d e = TxInputSpecimen {
-  txInputOut      :: TxOutSpecimen s d e,
-  txInputRedeemer :: RedeemerSpecimen s r}
-
-data TxMintSpecimen e = TxMintSpecimen {
-  txMintValue :: Value e}
-
-data WalletSpecimen e = WalletSpecimen {
-  walletPubKey :: PubKey}
-
-type TxOutSpecimen :: forall (s :: script) -> Datum s -> [currency] -> Type
-data TxOutSpecimen s d e = TxOutSpecimen {
-  txOutDatum :: DatumSpecimen s d,
-  txOutValue :: Value e}
-
-type Value :: [currency] -> Type
-data Value currencies = Value (AmountsOf currencies)
-
-type AmountsOf :: [c] -> Type
-data AmountsOf currencies where
-  Destitute :: AmountsOf '[]
-  (:$) :: Natural -> Proxy c -> AmountsOf '[c]
-  (:+) :: AmountsOf '[c] -> AmountsOf cs -> AmountsOf (c ': cs)
-
-infixr 3 :$
-infixr 2 :+
-
-data Ada = Ada
-
-data Collateral = CollateralAda
-
 type Wallet :: c -> k -> (c -> Type) -> Type
 data Wallet c s w
 
 type InputWallet :: c -> (forall s -> Redeemer s -> Datum s -> c -> Type) -> (c -> Type) -> Type
 data InputWallet c s w
+
+data Ada = Ada
